chg: move util functions to helpers
