+++ test/meta-tx/metaTx.test.js
@@ -9,7 +9,8 @@ import * as ethUtils from 'ethereumjs-util'
 
 import * as deployer from '../helpers/deployer'
 import { mockValues } from '../helpers/constants'
-import { generateFirstWallets, constructERC1155DepositData } from '../helpers/utils'
+import { generateFirstWallets, constructERC1155DepositData, getSignatureParameters } from '../helpers/utils'
+import { getTypedData } from '../helpers/meta-tx'
 import logDecoder from '../helpers/log-decoder.js'
 
 const ChildERC20 = artifacts.require('ChildERC20')
@@ -446,54 +447,3 @@ contract('NativeMetaTransaction', (accounts) => {
     })
   })
 })
-
-const getTypedData = ({ name, version, chainId, verifyingContract, nonce, from, functionSignature }) => {
-  return {
-    types: {
-      EIP712Domain: [{
-        name: 'name',
-        type: 'string'
-      }, {
-        name: 'version',
-        type: 'string'
-      }, {
-        name: 'chainId',
-        type: 'uint256'
-      }, {
-        name: 'verifyingContract',
-        type: 'address'
-      }],
-      MetaTransaction: [{
-        name: 'nonce',
-        type: 'uint256'
-      }, {
-        name: 'from',
-        type: 'address'
-      }, {
-        name: 'functionSignature',
-        type: 'bytes'
-      }]
-    },
-    domain: {
-      name,
-      version,
-      chainId,
-      verifyingContract
-    },
-    primaryType: 'MetaTransaction',
-    message: {
-      nonce,
-      from,
-      functionSignature
-    }
-  }
-}
-
-const getSignatureParameters = (signature) => {
-  const r = signature.slice(0, 66)
-  const s = '0x'.concat(signature.slice(66, 130))
-  const _v = '0x'.concat(signature.slice(130, 132))
-  let v = parseInt(_v)
-  if (![27, 28].includes(v)) v += 27
-  return { r, s, v }
-}
